Remove redundant util functions; move corresponding unit tests
